bookhawk-backend/requirements.txt
fastapi
uvicorn
python-dotenv
ultralytics
pillow
python-multipart
segment-anything-py
torch
torchvision